添加代办任务

+++ src-tauri/src/models/task_list.rs
@@ -1,5 +1,7 @@
-#[derive(serde::Deserialize, Debug)]
+use serde::Serialize;
+
+#[derive(serde::Deserialize, Debug, Serialize)]
 pub struct TaskList {
-    pub id: i32,
+    pub id: u64,
     pub name: String,
 }
